Finish the first implementation of the HomePageHandler
This is the first working implementation with the intention of keeping
the logic as simple as possible, while moving away anything not
closely related to the HomePageHandler to a service or other external
class. It does a reasonable job of checking for the required
prerequisites before moving on to carrying out the core processing of
the email, before, if able to, creating a note on the user's account.