Add http-status-code to provide constants for HTTP status codes
While, I feel, the HTTP status codes are pretty self-explanatory,
perhaps that's only my opinion. So, to account for that, for those new
to the topic, and for just general improved readability, this package is
being used to provide human-readable names for HTTP status codes in
place of using the integer values themselves.