Add and register factories to instantiate Twilio classes
This is a small change that creates and registers two classes with the DI
container to simplistically load the Twilio-related classes in the
application. I've deliberately kept them simple so as not to spend too
much time on them, which could be spent later, and to focus on other
aspects of the application.