Refactor UserNoteService::createNote to return a Note entity
The reason for this is to remove a second request to the EntityManager
to retrieve a User entity from the DB. The user's already retrieved in
createNote, it doesn't need to be instantiated again in HomePageHandler.
This change makes that happen, cleans up the code a bit, along with the
tests so that they stay in sync.